Escape double quotes with a double quote.
When exporting to CSV format, a double quote in the original text must be replaced with another double quote. If not, the CSV parser may see a semicolon and think that's a field separator.

Tested by loading in LibreOffice a CSV using semicolons to separate. We see that only a single double quote is shown, even though in the source and the output two sets of double quotes are shown. The format contains both a quoted word in the middle and also at the end of the field.

Issue: GNATS-Automation/637