Define the output coding to be ISO-8859-15.
When checking the import directly in DOORs, UTF8 characters are shown as ISO-8859-1/15. Because -15 is typical, this was chosen for the implementation. Should fix for the normal output as well as for the new plain text output.

Tested by doing an export and ensuring in Notepad++ that with ISO-8859-15 we see the correct character. Tested the old export that was UTF8 by configuring the encoding to ISO-8859-15 also and that the output is the same as in DOORs.

Issue: GNATS-Automation/637