Improve the robustness of the parsing by looking in the DOM instead of using flags.
Even though the previous commit worked, it took some time because of the InParagraph flag. This can be equivalently obtained by looking for a parent element of the name "para" and is always accurate (albiet may be a little bit slower). Maintenance over speed here.

Tested by ensuring an export has the same results as before.