Parse newlines within a text block into paragraphs.
Take a text block and look for line breaks in the form of #xA; and #xD;. When these are found, create a new paragraph block (ensuring not to create empty paragraph blocks).

Tested by exporting and comparing against EA and looks good, except for one case when a list is back-to-back.