Escape the CSV text after converting to the required encoding.
In special cases when we had curly quotes such as “HIGH”, this would be converted to single quotes by the .NET converters. But because the escaping occurred before the conversion, the resulting "HIGH" was not properly quoted. So we do away with the StreamWriter and use a normal stream, converting the data manually first and then writing directly to the stream.

This has the advantage that test cases don't need the special constructor and StreamWriter to flush the stream inside the using block.

Issue: GNATS/Automation-641