Read the Creation and Modified time from the XMI file.
It is now displayed on the main screen as two new fields. If the time is invalid, it is shown as 1/1/1970. If it is not defined, it is not shown at all. All times are assumed to be in UTC.

The DocBook 4.5 format exports this as part of each chapter.