Don't crash when importing XMI 1.1 with no Status field
The GUI would report problems when opening a XMI file that has an element with no 'status' attribute. Now a check is made to ensure that the status field is present or not, and is ignored if not present.

Export of data was checked and works as expected. DocBook will not provide status information if it wasn't in the original requirement.

Issue: DOTNET-153